### 2018-06-27 #793 @whedon
[PRE REVIEW]: effmass: An effective mass package
**Submitting author:** @lucydot (<a href="http://orcid.org/0000-0002-2992-9871">Lucy Whalley</a>)
**Repository:** <a href="https://github.com/lucydot/effmass" target ="_blank">https://github.com/lucydot/effmass</a>
**Version:** v1.0.b1
