### 2018-06-27 #793 @arfon
@lucydot - it looks like @whedon can't handle your colour selections 😸. Could you please remove this line from your `paper.md`?

```
 urlcolor: "TealBlue"
```
