### 2018-06-28 290694d
Remove teal blue colour for url links
