### 2018-06-28 #793 @lucydot
Hello @arfon and @katyhuff -

I have removed the offending line -

Perhaps @ajjackson might like to review the code. We worked in the same research group 2 years ago, but he has had no input on this project.

Looking at the reviewer database, the following people may also like to review:
@mdoucet
@bocklund
@mkhorton
@yxqd
