### 2018-06-28 #793 @katyhuff
@lucydot can you look through to deal with all colors that cause these errors? You can debug until it works with `@whedon generate pdf` here in the comments. Thanks!
