### 2018-06-28 #793 @katyhuff
Hi @ajjackson, @mdoucet, @bocklund, and @mkhorton - It seems to me you would be an excellent reviewer for this submission. As soon as it builds, are you willing to be assigned to review this software paper submission? We'll need at least two volunteers, but the more the merrier. Information about being a reviewer can be found here : http://joss.theoj.org/about#reviewer_guidelines .
