### 2018-06-28 fb34f14
Remove colourlink = True as it will not build with any colour
