### 2018-06-29 #797 @whedon
[REVIEW]: effmass: An effective mass package
**Submitting author:** @lucydot (<a href="http://orcid.org/0000-0002-2992-9871">Lucy Whalley</a>)
**Repository:** <a href="https://github.com/lucydot/effmass" target ="_blank">https://github.com/lucydot/effmass</a>
**Version:** v1.0.b1
**Reviewers:** @ajjackson @bocklund @@mkhorton

- [ ] **Version:** Does the release version given match the GitHub release (v1.0.b1)?

- [ ] **Version:** Does the release version given match the GitHub release (v1.0.b1)?

- [ ] **Version:** Does the release version given match the GitHub release (v1.0.b1)?
