### 2018-06-29 #793 @arfon
@katyhuff - if you want to add a second reviewer you need to do:

`assign` clobbers the existing assignments.
