### 2018-07-04 68d270f
from scipy import integrate - fixing import error
