### 2018-07-05 86299b8
added automatic push to pypi for tagged commits in travis yml
