+++ .travis.yml
@@ -1,12 +1,22 @@
 language: python
 python:
-  - "3.6"
-# command to install dependencies
+- '3.6'
 notifications:
   email: false
 install:
-  - pip install -r requirements.txt
-# command to run tests
-script: coverage run --rcfile=.coveragerc -m pytest 
+- pip install -r requirements.txt
+script: coverage run --rcfile=.coveragerc -m pytest
 after_success:
-  - codeclimate-test-reporter --token $CODECLIMATE_REPO_TOKEN
+- codeclimate-test-reporter --token $CODECLIMATE_REPO_TOKEN
+
+deploy:
+  provider: pypi
+  user: "lucydot"
+  on: 
+    tags: true
+    branch: master
+  password:
+    secure: XcFD/nlL4hyV+gg0eFibx9QML9bJSQiRgNQmdHXz/v6ZT0QmYGDvHU6MOxs2/RNYjcfKh5QjPoyrXyNjIu9urz4MeGVrwLzh/aA3ZGWeVP+jIHEAClYIvkqSFQvrWTngd96KE9RxQUSRFq8Vm8icSoZx3i0lEH3ZxX/hN5OG0K26CaUiid2f/W6MugHZoY9SEDlZpLDq1otd4fwv0Odkqxq6lixfbpymrmmAr/O3xt3QcD+pQYKHmW595AbqA/AepfGF7RHYeUfMB6mO1ApVD1GMSBQg4JdqwGuwqOs5T7yAX7la4+563qF2f77V6bbvg1IXhvwT/wpXgznUaF0UFbBcVZAcQ+eZGUXVvjp2EzMhMFl//0hygTfmL/4GA7/ZTHLBPiwhiR07QYBgM9Rytl6QPV9USPpjSVpNNa/cAp/hzlgtVeZbeI7sATk18iK4MfK7Xu51whgm/AD/OplIGLjxGV8k2qAHnhL0iDONPeq0eh6Ec3NWCb0Bj8MkiwalnbndMAwE7Ct15/8m6Z3wgLqX1fwSJW5gSdTyhbyF/yC/y+PfwgSpCLXZEqta8g/yRNZp6XtR/OwmXF3WkIW9rtRRXazIqPIMBLfsb3VH5mM7DRXpnB8Aft4n0dPmkRPSlwmJwtNDMk6iDdu6P5BIHvAMoh+JEPNSIq+urDO3a2M=
+env:
+  global:
[…]
