### 2018-07-05 7beda61
attempt at fix for authenticating automatic pypi upload
