+++ .travis.yml
@@ -12,11 +12,7 @@ after_success:
 deploy:
   provider: pypi
   user: "lucydot"
+  secure: "MUTzcBaV0uy/1fDdiXhKsfdEx2dM/tiTlWk0sxBAmo6v+i4eamKDkMsuVtnuq8cOY219Y/8W4dqeXsuFKnLOSvsFGMZ6G6KqmcGT4budWhP7AAPJW9586MZrw6Ib7igFkNsnJGiL9CUO9DY0i+ZyZB7lHx1Co7BT+bkBNQCP+XxkVgWpVavSTN1rnLefws2IIaYUDmO/Da4EHK4jz0IM1UcftuUGf677imABTupHSh0U2MCywwkpQgiJUStP5xZBeBIPQ0IxFIU4IU2BCsUZNirEnzvk7FegPHZ35jgQ1IfsTP7pFr2SuTd+D4fKbn/pEaSL/uAUKI1vaLJmwVAgv8EOJAqePMR81Aemlk+8aDR+jSXICSdN9AgEwpElyektAUSY/ud4FZE7x0kcn7GCLjiNlADEUXyp1RsgNtGaqprlhXaURtaRD66pnNA3TE4DG++LQr1x7111pX9hxexqZAMxur8M1xijLHNMASoKf+bQhlsqZLsMGbAAw94wO4KufkentAdQfhnvzoHSuW2wDXyDmXo5r7h7XzLj2HNgddztF8/6m3rPGZ/HlPX2sNDFqHNSwbb4RWCu51XImiqd5Ov2puDr/b1C+UHz/ZRQofXRuZFc8rH/E8Rff95MW6xhYDw9kyUuqG3o/3iPUjkp3/VT16YVgFTyMMAqmucOZm4="
   on: 
     tags: true
     branch: master
-  password:
-    secure: XcFD/nlL4hyV+gg0eFibx9QML9bJSQiRgNQmdHXz/v6ZT0QmYGDvHU6MOxs2/RNYjcfKh5QjPoyrXyNjIu9urz4MeGVrwLzh/aA3ZGWeVP+jIHEAClYIvkqSFQvrWTngd96KE9RxQUSRFq8Vm8icSoZx3i0lEH3ZxX/hN5OG0K26CaUiid2f/W6MugHZoY9SEDlZpLDq1otd4fwv0Odkqxq6lixfbpymrmmAr/O3xt3QcD+pQYKHmW595AbqA/AepfGF7RHYeUfMB6mO1ApVD1GMSBQg4JdqwGuwqOs5T7yAX7la4+563qF2f77V6bbvg1IXhvwT/wpXgznUaF0UFbBcVZAcQ+eZGUXVvjp2EzMhMFl//0hygTfmL/4GA7/ZTHLBPiwhiR07QYBgM9Rytl6QPV9USPpjSVpNNa/cAp/hzlgtVeZbeI7sATk18iK4MfK7Xu51whgm/AD/OplIGLjxGV8k2qAHnhL0iDONPeq0eh6Ec3NWCb0Bj8MkiwalnbndMAwE7Ct15/8m6Z3wgLqX1fwSJW5gSdTyhbyF/yC/y+PfwgSpCLXZEqta8g/yRNZp6XtR/OwmXF3WkIW9rtRRXazIqPIMBLfsb3VH5mM7DRXpnB8Aft4n0dPmkRPSlwmJwtNDMk6iDdu6P5BIHvAMoh+JEPNSIq+urDO3a2M=
-env:
-  global:
[…]
