### 2018-07-05 83a6731
another attempy at travis yml pypi integration
