+++ .travis.yml
@@ -12,7 +12,8 @@ after_success:
 deploy:
   provider: pypi
   user: "lucydot"
-  secure: "MUTzcBaV0uy/1fDdiXhKsfdEx2dM/tiTlWk0sxBAmo6v+i4eamKDkMsuVtnuq8cOY219Y/8W4dqeXsuFKnLOSvsFGMZ6G6KqmcGT4budWhP7AAPJW9586MZrw6Ib7igFkNsnJGiL9CUO9DY0i+ZyZB7lHx1Co7BT+bkBNQCP+XxkVgWpVavSTN1rnLefws2IIaYUDmO/Da4EHK4jz0IM1UcftuUGf677imABTupHSh0U2MCywwkpQgiJUStP5xZBeBIPQ0IxFIU4IU2BCsUZNirEnzvk7FegPHZ35jgQ1IfsTP7pFr2SuTd+D4fKbn/pEaSL/uAUKI1vaLJmwVAgv8EOJAqePMR81Aemlk+8aDR+jSXICSdN9AgEwpElyektAUSY/ud4FZE7x0kcn7GCLjiNlADEUXyp1RsgNtGaqprlhXaURtaRD66pnNA3TE4DG++LQr1x7111pX9hxexqZAMxur8M1xijLHNMASoKf+bQhlsqZLsMGbAAw94wO4KufkentAdQfhnvzoHSuW2wDXyDmXo5r7h7XzLj2HNgddztF8/6m3rPGZ/HlPX2sNDFqHNSwbb4RWCu51XImiqd5Ov2puDr/b1C+UHz/ZRQofXRuZFc8rH/E8Rff95MW6xhYDw9kyUuqG3o/3iPUjkp3/VT16YVgFTyMMAqmucOZm4="
+  password:
+    secure: "MUTzcBaV0uy/1fDdiXhKsfdEx2dM/tiTlWk0sxBAmo6v+i4eamKDkMsuVtnuq8cOY219Y/8W4dqeXsuFKnLOSvsFGMZ6G6KqmcGT4budWhP7AAPJW9586MZrw6Ib7igFkNsnJGiL9CUO9DY0i+ZyZB7lHx1Co7BT+bkBNQCP+XxkVgWpVavSTN1rnLefws2IIaYUDmO/Da4EHK4jz0IM1UcftuUGf677imABTupHSh0U2MCywwkpQgiJUStP5xZBeBIPQ0IxFIU4IU2BCsUZNirEnzvk7FegPHZ35jgQ1IfsTP7pFr2SuTd+D4fKbn/pEaSL/uAUKI1vaLJmwVAgv8EOJAqePMR81Aemlk+8aDR+jSXICSdN9AgEwpElyektAUSY/ud4FZE7x0kcn7GCLjiNlADEUXyp1RsgNtGaqprlhXaURtaRD66pnNA3TE4DG++LQr1x7111pX9hxexqZAMxur8M1xijLHNMASoKf+bQhlsqZLsMGbAAw94wO4KufkentAdQfhnvzoHSuW2wDXyDmXo5r7h7XzLj2HNgddztF8/6m3rPGZ/HlPX2sNDFqHNSwbb4RWCu51XImiqd5Ov2puDr/b1C+UHz/ZRQofXRuZFc8rH/E8Rff95MW6xhYDw9kyUuqG3o/3iPUjkp3/VT16YVgFTyMMAqmucOZm4="
   on: 
     tags: true
     branch: master
