### 2018-07-05 9b11220
edits for pypi integration to yml
