### 2018-07-05 3c7160f
edits for pypi integration to yml
