### 2018-07-06 lucydot/effmass#1 @bocklund
SciPy is not imported, leading to errors in printing results
`effmass/analysis.py` imports `integrate` by name, however the `mass_integration` function uses the qualified `scipy.integrate.quad`. This leads to an error in line `outputs.print_results(segments[-1], data, settings)`, which fails with the following traceback:

```python
NameError                                 Traceback (most recent call last)
<ipython-input-26-204cc925e032> in <module>()
----> 1 outputs.print_results(segments[-1], data, settings)

~/Projects/joss-reviews/effmass/effmass/outputs.py in print_results(segment, data, settings, polyfit_order)
    136                   polyfit_order=polyfit_order)]))
    137     print("optical mass at band edge (assuming the Kane dispersion) is {:.2f}".
--> 138           format(segment.optical_effmass_kane_dispersion()))
    139
    140     plt.figure(figsize=(8, 8))

~/Projects/joss-reviews/effmass/effmass/analysis.py in optical_effmass_kane_dispersion(self, fermi_level, temp, alpha, mass_bandedge, upper_limit)
    488             alpha=alpha,
    489             mass_bandedge=mass_bandedge,
--> 490             upper_limit=upper_limit)
    491         bottom = self.weight_integration(
    492             fermi_level=fermi_level,

~/Projects/joss-reviews/effmass/effmass/analysis.py in mass_integration(self, fermi_level, temp, alpha, mass_bandedge, upper_limit)
    334         fermi_level = self.fermi_energy if fermi_level is None else fermi_level
    335
--> 336         result = scipy.integrate.quad(
    337             self._mass_integrand,
    338             0,

NameError: name 'scipy' is not defined
```

Fixing `scipy.integrate.quad` -> `integrate.quad` should fix this.

I also suggest writing a test for this code path.

(Part of https://github.com/openjournals/joss-reviews/issues/797)
