### 2018-07-06 lucydot/effmass#1 @bocklund
Also affects the lines where optical mass is calculated in the Jupyter Notebook examples, e.g. `optical_mass_111 = [segments[-6].optical_effmass_kane_dispersion(fermi_level=data.CBM+burstein_moss_kane(x,average_mass,average_alpha),alpha=0.16/ev_to_hartree,mass_bandedge=0.155,upper_limit=0.1) for x in concentrations]`
