### 2018-07-06 #797 @bocklund
Here is my first pass:

### General checks

- **Version:** The latest PyPI version is 0.1, behind 0.1.1 and 1.0.b1, but it seems like the the author is working on CI-backed releases. The released version does not work with the current examples on GitHub.

### Functionality

- **Functionality:**  Issue https://github.com/lucydot/effmass/issues/1 was opened. Once closed this can be checked off.

### Documentation

- **A statement of need:** effmass depends on pymatgen through vasppy. @lucydot, can consider comparing effmass to existing software like pymatgen or ASE (what it provides that those do not) in the documentation?
- **Community guidelines:** There are not very clear guidelines for contributing or a code of conduct (item 1). Support seems to implicitly offered through GitHub, but maybe this is not the author's intention (item 3).

### Other comments

- Since this already depends on pymatgen, is it possible that effmass can support parsing electronic structures from other software?
