### 2018-07-09 e55f7f9
scipy.integrate --> integrate as imported integrate by name
