### 2018-07-09 3d006b7
Outline which features are unique to the effmass package
