### 2018-07-09 eccf1c5
Update Features.rst
