### 2018-07-09 947cfd1
Create CODE_OF_CONDUCT.md
