### 2018-07-09 fbdfdf4
Include code of conduct in development section
Also suggest possible contribution, parsing of other electronic structure codes.
