### 2018-07-09 0de1022
added test for the optical_effmass_kane_dispersion function
