### 2018-07-09 ae50426
highlighted the need to account for the sign of the alpha parameter and kane mass at band edge in the documentation
