### 2018-07-09 6d24e38
Comment out automatic deployments to PyPI
A task for another day...
