+++ .travis.yml
@@ -8,14 +8,14 @@ install:
 script: coverage run --rcfile=.coveragerc -m pytest
 after_success:
 - codeclimate-test-reporter --token $CODECLIMATE_REPO_TOKEN
-deploy:
-  provider: pypi
-  user: lucydot
-  password:
-    secure: "WExlC74LSaofcGZi9d+mZBpspFNfBA961q2gRbSGKOJLjNbszmw6AOpu4A2QSqHuQmbRX322AOQ9i3iFC6tCI0zAkKyb6q3DHj53qp8QPAR/oNQjPDwHg3W4u1+ZOvxShgsc7XViSyNuZCa0Y9kWjvH4BHtsNDeQJmqOJlsnxYr2SD+xMeoWvEi3xZ1ysT3HOqWYHBcHnc4VzaqnJhNlwYyFCmdE3ded8ES8gKLHqz5ZvQa1gsfSdSiDNfgeStCo9WVy6kHdKtWuLB/7CH37C4DFglAO41MW+BmIQckXZKZGDpsnb41q11rHsAqiVP+wm6Hfx3EQQzJJjF8C2pTBvaT/urs/QFEM5RQNzRtFemABikvN6153W5pQFBNn9RCjtJJDj0sHbRNCiDOQQ1EhTS0HL4bF7rfxyKWUFDwXGy0BXUk7SUzxPEogJueml9r/43W0Y+2bEcAxmpATPrWyCTQ9Evbxq0EjibbM/900n23Ysf5zwIpgpEasGfCDLSvNPhYGfjh7vePlnoT60zfuecCzWuzGdEHxt93yV17bk9yOuBHlI7v5gmEs4gp+CaqeKIz9Vi6QorR+mc2dN6GRMs1e+SlszpQ+LrPKVL6z+lZNDLwVkIo5mQlsz1x1FfFnkfyaGX+3CqfkD7O+btBYkLBSJBLGxZrl9ydsYuQ2PMo="
-  on:  
-    tags: true
-    branch: master
-env:
-  global:
[…]
