### 2018-07-09 211f325
Update readthedocs to reflect latest changes to README
