+++ docs/source/Development.rst
@@ -2,6 +2,6 @@
 Development
 ===========
 
-Please use the Github [issue tracker](https://github.com/lucydot/effmass/issues/) for feature requests and bug reports. 
+Please use the Github `issue tracker <https://github.com/lucydot/effmass/issues/>`_ for feature requests and bug reports. 
 
-If you would like to contribute please do so via a pull request. All contributors must read and respect the [code of conduct](https://github.com/lucydot/effmass/blob/master/CODE_OF_CONDUCT.md). In particular, we welcome contributions which would extend `effmass` so that it is able to parse output from other electronic structure codes. 
+If you would like to contribute please do so via a pull request. All contributors must read and respect the `code of conduct <https://github.com/lucydot/effmass/blob/master/CODE_OF_CONDUCT.md>`_. In particular, we welcome contributions which would extend `effmass` so that it is able to parse output from other electronic structure codes. 
