+++ README.md
@@ -28,7 +28,7 @@ These correspond to the valence band maxima and conduction band minima. Maxima a
 The curvature (aka inertial), transport and optical effective masses are calculated using the derivatives of a fitted polynomial function. The optical effective mass can also be calculated assuming a Kane dispersion.
 
 **Assess the extent of non-parabolicity:**
-Parameters of the Kane quasi-linear dispersion are calculated to quantify the extent of non-parabolicity over a given energy range. add
+Parameters of the Kane quasi-linear dispersion are calculated to quantify the extent of non-parabolicity over a given energy range. 
 
 **Calculate the quasi-fermi level for a given carrier concentration:**
 This requires the `VASP` output file `DOSCAR`. Using density-of-states data and assuming no thermal smearing, `effmass` can calculate the energy to which states are occupied. This is a useful approximation to the quasi-Fermi level.
