### 2018-07-09 a4658d1
bump version number for upload to PYPI
