### 2018-07-09 lucydot/effmass#1 @lucydot
Thank you for finding this bug @bocklund. I have fixed this as you suggested and have written [a test](https://github.com/lucydot/effmass/blob/a4658d1c200eb3f0f02550efe868d2b4d4b294fc/tests/test_analysis.py#L99-L102)  to catch similar failures in the future.
