### 2018-07-09 #797 @lucydot
Hi @bocklund and thank you for taking the time to look over `effmass`

### General checks

- **version**: The latest release contains the changes below and I have uploaded this to PyPI; it should install with `pip install effmass=1.0b3`

### Functionality

- **Functionality**: I have fixed the bug and closed the issue: please see lucydot/effmass#1 for details

### Documentation

- **A statement of need**: Thank you for this idea - I have added a paragraph to the documentation [here](https://effmass.readthedocs.io/en/latest/Features.html).
- **Community guidelines**: I had not thought of including a code of conduct but I think it is a great idea. I have added  [`CODE_OF_CONDUCT.md`](https://github.com/lucydot/effmass/blob/master/CODE_OF_CONDUCT.md) and link to it from the [`README.md`](https://github.com/lucydot/effmass/blob/master/README.md)

### Other comments

- I agree that it would be nice for effmass to support parsing of electronic structures from other software, and it should be relatively easy to implement. I aim to set time aside for a future release, and have suggested it as a possible contribution in the [`README.md`](https://github.com/lucydot/effmass/blob/master/README.md).
