### 2018-07-09 #797 @ajjackson
Yes, I have been looking at it today!

As Lucy mentioned in the Pre-review thread, we were members of the same research group a little over two years ago but we never published together. I was not aware of the existence of this code until asked to review it.

I will check off what I can now and then try out the code on some fresh data.
