### 2018-07-09 #797 @ajjackson
As a developer of Sumo (which is also going through JOSS review), I am comfortable with the amount of overlap between effmass and Sumo. Effmass is more flexible for effective mass fitting and concentrates on providing a toolkit for analysis. Sumo is a suite of command-line tools; it is more opinionated and "fire-and-forget", with a less Python-adept target audience.
