### 2018-07-09 #797 @ajjackson
In general **effmass** seems to be packaged and documented very nicely. I need to take a bit more time to play with the actual functionality. Here are some comments in the meantime:

## General checks
- **version**: the PyPI/beta/source versioning was a touch confusing but as far as I can see the packaging all works as it should and I was able to get the 1.01 beta versions easily enough. I think this is ok.

## Documentation
- **Statement of need**: The summary of unique features is helpful, a very useful thing to have when people are looking for tools. I would like to see a more explicit statement somewhere of the target audience (i.e. theoretical solid-state physics/chemistry community). The review guidelines imply that this should be in both the paper and other docs.
- **Installation instructions**:
  - The Python package management works well and I don't think a separate requirements list is essential. I do think it is worth mentioning that Numpy/Matplotlib are used as these dependencies are a) quite large b) sometimes best installed with a different package manager.
  - Numpy is listed as a dependency in setup.py and requirements.txt, but SciPy is not. Scipy is imported in `analysis.py` so seems to be a required module. It _is_ one of Pymatgen's many dependencies so the installers work, but this seems inconsistent when other packages installed by Pymatgen are listed explicity.
  - requirements.txt is missing Pymatgen. I assume this is related to making readthedocs work, as we ran into exactly this issue on Sumo. Probably the best solution is to name it requirements_rtd or something so that people don't try to use it with virtualenv.
- **Community guidelines:** The guidelines for contributors are good. The guidelines for bugs/support are minimal but do the job.

@bocklund and @mkhorton have both noted the Pymatgen dependency as a basis for further development. Poking around though, I'm not sure that effmass actually uses Pymatgen at all? Perhaps I've missed something, but as far as I can see it's just a side-effect of installing Vasppy?
