### 2018-07-09 #797 @mkhorton
@ajjackson yeah, it's a little weird -- I'm not actually familiar with vasppy, but it looks like pymatgen is a dependency of vasppy but it doesn't actually use pymatgen much. In particular, vasppy itself doesn't seem to build on pymatgen's VASP classes, but rather has its own VASP classes, which means building in pymatgen support to effmass would indeed be a bit trickier. (Hello by the way, it's a small world -- I didn't realise you were the developer of sumo!)
