### 2018-07-11 3f6c7aa
Impove installation instructions
- Highlight dependancy on SciPy ecosystem
- pip-->pip3 and python-->python3 as some may get python2 and python3 installs mixed up
