### 2018-07-11 eeefaf8
Add scipy
This was being installed as dependancy of vasppy (via pymatgen) but it should be listed explicitly
