### 2018-07-11 2cb18d9
include a section about related packages and unique features to the paper
