+++ paper.md
@@ -36,6 +36,32 @@ It contains a core class that calculates the effective mass and other associated
 [``effmass``](https://github.com/lucydot/effmass) also contains functions for locating band structure extrema, calculating the Kane quasi-linear dispersion parameters and plotting approximations to the true dispersion.
 Parsing of electronic structure data is faciliated by the [``vasppy``](https://github.com/bjmorgan/vasppy) [@Morgan2018] package.
 
+The `effmass` package is aimed towards theoretical solid state physicists and chemists who have a basic familiriarity with Python. If you are looking to calculate the parabolic effective mass only, and would prefer a command line interface, it may be that one of the packages listed below will suit your needs better.
+
+# Related packages
+
+Effective mass calculations are implemented in a number of other packages:
+
+[vasppy](https://github.com/bjmorgan/vasppy/): This is installed as a dependancy of `effmass`. Calculates the effective mass using a least-squares quadratic fit for parabolic dispersions. 
+
+[sumo](https://github.com/SMTG-UCL/sumo): Calculates the effective mass using a least-squares fit for parabolic and non-parabolic dispersions. 
+
+[emc](https://github.com/afonari/emc): Calculates the effective mass *tensor* using a finite-difference method for parabolic dispersions.
+
+[pymatgen](http://pymatgen.org/): This is installed as a dependancy of `effmass`. Calculates an average effective mass *tensor* for non-parabolic dispersions with multiple bands and extrema. Also calculates the Seebeck effective mass as defined [here](https://perso.uclouvain.be/geoffroy.hautier/wp-content/papercite-data/pdf/gibbs2017.pdf).
+
+If you have an update to the information above then please use the Github [issue tracker](https://github.com/lucydot/effmass/issues/). 
+
+## Which features are unique to the effmass package? 
+
+To our knowledge, the following features are unique to this package:
+
+- easily compare the values of curvature effective mass calculated using multiple numerical techniques (least-squares and polynomial fitting)
+- tailor the polynomial fitting used to approximate the DFT calculated dispersion: by choosing the order of the polynomial and the energy range to fit over.
+- visualise the dispersions used to approximate the DFT calculated dispersion
+- quantify non-parabolicity through the Kane dispersion parameters: effective mass at band-edge and alpha
+- calculate the optical effective mass assuming a Kane dispersion.
+
 # Acknowledgements
 
 LW would like to thank [Aron Walsh](https://github.com/aronwalsh), [Benjamin Morgan](https://github.com/bjmorgan) and [Jarvist Moore Frost](https://github.com/jarvist) for their guidance during this project.
