### 2018-07-11 892cf7d
improve wording in paper
