+++ docs/source/Installation.rst
@@ -2,16 +2,19 @@
 Installation
 ============
 
+`effmass` is a Python 3 package and requires key packages from the `SciPy ecosystem <https://www.scipy.org/about.html>`_: SciPy, NumPy and Matplotlib. If you have not installed these packages before, it may be best to install them using your preferred package manager (eg: Homebrew). Note that together they will use >100MB of disk space. `effmass` can then be built using the Python package manager `pip`:
+
+
 .. code-block:: none
 
-   pip install effmass
+   pip3 install --user effmass
 
 Or download the latest release from `GitHub <https://github.com/lucydot/effmass/releases>`_, and install
 
 .. code-block:: none
 
    cd effmass
-   python setup.py install
+   python3 setup.py install
 
 Or clone the latest development version
 
@@ -24,4 +27,4 @@ and install the same way.
 .. code-block:: none
 
    cd effmass
-   python setup.py install 
+   python3 setup.py install 
