+++ setup.py
@@ -21,7 +21,8 @@ config = {
     'download_url': "https://github.com/lucydot/effmass/archive/%s.tar.gz" % (__version__),
     'version': __version__,
     'install_requires': [ 'vasppy', 
-			              'numpy', 
+			  'scipy',
+			  'numpy', 
                           'matplotlib',
                           'codeclimate-test-reporter',
                           'pytest-lazy-fixture',
