### 2018-07-11 #797 @lucydot
Thank you @ajjackson and @mkhorton for your suggestions.

### Overlap with other packages
I've added a `Related packages` section to the [paper.md](https://github.com/lucydot/effmass/blob/master/paper.md). This has also been added as a [new page in RTD](https://effmass.readthedocs.io/en/latest/Related%20packages.html), and a link to this page has been added to the [README](https://github.com/lucydot/effmass/blob/master/README.md).

### Target audience
I have explicitly stated the target audience as part of the `Related packages` documentation (see above).

### Installation instructions
I have included explicit mention of the scipy, numpy and matplotlib packages as suggested.

### Dependencies: scipy
Thank-you for pointing out I had missed listing scipy as a dependency in in requirements.txt and setup.py - this was an oversight and it is now added.

### Dependancies: pymatgen
I haven't included pymatgen as effmass does not make use of this (it is only installed as it is a dependency of vasppy).  To enable easy extension to other electronic structure code outputs, there is definitely an argument for losing the vasppy dependency and using pymatgen instead, though this would involve re-writing a significant part of the code.
