### 2018-07-13 #797 @katyhuff
Thank you @lucydot . It looks like you've addressed the reviewer comments. Let's see if they can recommend publication.

@mkhorton & @ajjackson : Thank you for your reviews. Do you have any remaining comments? Are you ready to make a final recommendation for acceptance based on this response, or would you like more time to explore the package?
