### 2018-07-13 #797 @mkhorton
I've updated my checklist. Ideally, I'd like some further time to play with this to check off the remaining items, but I'm happy to defer to the other referees if they're happy to accept now -- it certainly seems like a very nicely structured package and will be a valuable addition to the community.
