### 2018-07-14 lucydot/effmass#3 @ajjackson
Feature request: Printable representation of Segment
It would be nice if Segment objects had a standard printable representation, such as `2.3 eV [0 0 0] -> [1 0 1]`. This can can be implemented by providing a `__str__` method. It would make it easier to keep track of them and work with them interactively. The method implied in the tutorial is to repeatedly call `plot_segments` until you're sure you have the index of the one you want.
