### 2018-07-14 lucydot/effmass#4 @ajjackson
Broken link to data directory
It wasn't too hard to find the sample data under paper/data, but the link from the Jupyter notebook is broken for me. I think it should be https://github.com/lucydot/effmass/tree/master/paper/data ?

http://github.com/lucydot/effmass/paper/data gets me a 404.
