### 2018-07-14 #797 @ajjackson
I have worked through the Jupyter tutorial and was able to follow it and reproduce the plots without any trouble. It gives an applied overview of features; the comparison with DOS filling is a nice touch.

- [Issue raised](https://github.com/lucydot/effmass/issues/4) over broken link to data
- I can reproduce step [5] fine but it's not clear how it should be interpreted. The docstring says _"Plots bandstructure with the DFT-calculated points for each Segment instance overlaid."_ but if I had to guess what the vector labels are I'd say they are directions? It would be really helpful if these labels also included the segment indices.
- I have raised a feature request [here](https://github.com/lucydot/effmass/issues/3) which I think would improve usability. This should not be considered a condition of acceptance, it's a "nice to have".

I'd like to throw some of my own data into the mix before signing off
