### 2018-07-16 8478330
print index on segment and add string method to segment class
