### 2018-07-16 cdb7ce5
improve formatting of Segment string method
