+++ effmass/analysis.py
@@ -114,8 +114,16 @@ class Segment:
         self._VBM = Data.VBM
 
     def __str__(self):
+        """
+        Segment string method.
+
+        Returns:
+            A string containing the energy of the Segment extrema (referenced to the VBM) and the start- and end- points of the Segment in reciprocal space.
+        """
         energy_str = "{0:.2f}".format(self.energies[0] - self._VBM)
-        return energy_str+" eV;"+str(self.kpoints[0])+"-->"+str(self.kpoints[-1])
+        start_str = str(np.round(self.kpoints[0],3))
+        end_str = str(np.round(self.kpoints[-1],3))
+        return energy_str+" eV; "+start_str+"-->"+end_str
 
     def _check_kanefit_points(self, polyfit_order=6):
         """Raises an AssertionError if there are not enough data points to fit
