### 2018-07-16 c600eea
use adjustText for clever placement of text on plot
