### 2018-07-16 lucydot/effmass#3 @lucydot
Thanks for the suggestion - I've implemented this feature [here](https://github.com/lucydot/effmass/blob/55bfaf9d46d62fd9462dec03663d72cc6b4253b8/effmass/analysis.py#L116-L126) and a usage example is [in the notebook](https://github.com/lucydot/effmass/blob/master/paper/notebook.ipynb) (under 'Generate band segments').
