### 2018-07-16 #797 @lucydot
Thanks for the feedback @ajjackson:

- I have fixed the link and closed the issue.
- The labels now include the segments argument index. See [the notebook](https://github.com/lucydot/effmass/blob/master/paper/notebook.ipynb) for the example output - have I interpreted your suggestion correctly? The labels tended to overlap, so I have used the adjustText package to avoid this, and I think the output is much easier to read now. The function docstring and notebook have been updated so that it is clear what the labels correspond to.
- I have implemented the feature request: https://github.com/lucydot/effmass/issues/3 - thanks for the suggestion, it echos a similar (verbal) suggestion from @Guaxixi .
