### 2018-07-16 lucydot/effmass#5 @ajjackson
Partial occupancy problems
The code seems to be quite sensitive to partial occupancies. My first attempt with a poorly-converged calculation raised some `vasppy` errors; I think it didn't like occupancy "-0.0000" but didn't quite get to the bottom of it. A healthier calculation with some slight partial occupancy is giving me:

```
partial occupancy, particle type unknown. Please set Segment.ptype manually.
cannot determine bandedge energy as particle type unknown. Please set Segment.ptype.
```

How is a user meant to proceed from there? The docstring for a Segment gives
```
ptype (str): The quasi particle type, determined by occupancy of the eigenstate.
```
but there isn't a list of valid values.

It would be nice to either have some guidance for this case or a big warning that partial occupancies are not tolerated. It can be hard to avoid them, though.
