### 2018-07-21 #797 @ajjackson
Updates:

**Functionality documentation:** Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?
In general effmass is following best practice in the use of Python docstrings and automatically-generated documentation. I have checked this off as satisfactory, with the following observations:
- I found a minor error in a docstring, raised as an issue here: https://github.com/lucydot/effmass/issues/6
- It would be nice to have a bit more description in the API documentation re: what the modules are for. My own code documentation is also sparse in this respect... something to thing about for the future!

** References** It's difficult to check the references since the paper was updated as the PDF manuscript hasn't been rebuilt. I don't see a formal academic citation for the Seebeck effective mass paper (is that deliberate to indicate that it isn't used in this package?), or for pymatgen (https://doi.org/10.1016/j.commatsci.2012.10.028).

**Functionality** My attempt to test with independent data was foiled by some issues with partial occupancies; I will have another go before signing off if that's ok. https://github.com/lucydot/effmass/issues/5
