### 2018-07-21 #797 @katyhuff
@ajjackson please see built pdf above (to see what whedon can do, you can execute `@whedon commands` at any time)
