### 2018-07-21 #797 @katyhuff
@ajjackson  regarding your concern about api docs, the readme for the repo links to these API docs which seem quite sufficient to me : https://effmass.readthedocs.io/en/latest/
