### 2018-07-21 #797 @ajjackson
I agree they are sufficient, it's not a big deal  :-) I was referring to pages such as https://effmass.readthedocs.io/en/latest/API%20documentation.html which currently contains links without descriptions. This is very common in automatically generated documentation.
