### 2018-07-21 #797 @katyhuff
Ah, I misunderstood. Your comment initially made me think you couldn't find your way to the API pages.

Yes, I agree this is the default sphinx behavior. More might be better on that page, perhaps. But in this case, I think the docs are strong and it's not necessary if the reader has the patience to click through the links.
