### 2018-07-23 403ffe7
fixed docstring for generate_segments
