### 2018-07-23 lucydot/effmass#6 @lucydot
Thanks @ajjackson I have updated the docstring [here](https://github.com/lucydot/effmass/blob/403ffe7ff2a14c6d6f1849f38e77ccaabbd63e69/effmass/extrema.py#L332-L342).
