+++ paper.bib
@@ -49,4 +49,17 @@ issn = "0022-3697",
 url = "https://doi.org/10.1016/0022-3697(57)90013-6",
 author = {{E. O. Kane}}
 }
+
+@article{Ong2013,
+title = "Python Materials Genomics (pymatgen): A robust, open-source python library for materials analysis",
+journal = "Computational Materials Science",
+volume = "68",
+pages = "314 - 319",
+year = "2013",
+issn = "0927-0256",
+doi = "https://doi.org/10.1016/j.commatsci.2012.10.028",
+url = "http://www.sciencedirect.com/science/article/pii/S0927025612006295",
+author = "Shyue Ping Ong and William Davidson Richards and Anubhav Jain and Geoffroy Hautier and Michael Kocher and Shreyas Cholia and Dan Gunter and Vincent L. Chevrier and Kristin A. Persson and Gerbrand Ceder",
+keywords = "Materials, Project, Design, Thermodynamics, High-throughput"
+}
 	
