### 2018-07-23 b358982
added module level docstrings
