+++ docs/source/API documentation.rst
@@ -6,7 +6,33 @@ API documentation
    :maxdepth: 1
 
    inputs
+
+.. automodule:: effmass.inputs 
+
+.. toctree::
+   :maxdepth: 1
+
    extrema
+
+.. automodule:: effmass.extrema 
+
+.. toctree::
+   :maxdepth: 1 
+
    dos
+
+.. automodule:: effmass.dos 
+
+.. toctree::
+   :maxdepth: 1
+
    analysis
+
+.. automodule:: effmass.analysis 
+
+.. toctree::
+   :maxdepth: 1
+   
    outputs
+
+.. automodule:: effmass.outputs 
