### 2018-07-23 lucydot/effmass#7 @ajjackson
Error loading a spin-polarised calculation
I ran an LDA band structure for MgO. With no spin enabled it reads in ok, but when I set ISPIN=2 and recalculated the band structure effmass seems to have trouble reading the files.

Are spin-polarized calculations supported? I see that `effmass.inputs.Data` has an attribute for the number of spin channels, but I get an error while the object is being instantiated.

[spin_test.zip](https://github.com/lucydot/effmass/files/2220462/spin_test.zip)
