### 2018-07-23 #797 @mkhorton
@lucydot running `python setup.py test` fails with:

```
  File ".../effmass/tests/test_inputs.py", line 10, in <module>
    (pytest.lazy_fixture('MAPI_settings_object'), 0.025, 0.25),
AttributeError: module 'pytest' has no attribute 'lazy_fixture'
```

Your tests are working correctly with `python -m pytest` as recommended by your README, so this is not a bug, but I thought I'd bring this to your attention since it might trip people up.
