### 2018-07-24 #797 @ajjackson
I've raised an issue [here](https://github.com/lucydot/effmass/issues/7) about spin-polarised calculations, happy to tick the last box once that's resolved!

I intend to use this package.  It provides convenient comparison of different fitting methods, which is a good sanity check, and implements a superior weighted method for effective-mass determination. The Kane model implementation provides a useful way of quantifying non-parabolicity. The packaging follows best practices, and decent API documentation is accompanied by a useful tutorial.
