### 2018-07-24 89fbdc2
change ptype / electron / hole to band_type / valence_band / conduction_band
