### 2018-07-24 b3b93a8
change 0.5 to cutoff for whether considered occupied or not
