### 2018-07-24 887f19b
improve error handling for cases where the fermi-dirac distribution is less than python floating point precision
