### 2018-07-24 cd31687
error handling for the summary output
