### 2018-07-24 2b640a9
update tests for recent change to variable names
