### 2018-07-24 4f08757
clarify error message for partial occupancy in bandedge function
