### 2018-07-24 9280b9c
update tests with new variable names
