### 2018-07-24 204c035
AssertionError ---> ValueError
