### 2018-07-24 lucydot/effmass#5 @lucydot
I've improved the error handling for partial occupancies so that it is clearer which steps to take [here](https://github.com/lucydot/effmass/blob/2b640a9d2cbae3474b61425ba52a090152012f44/effmass/analysis.py#L193-L194) and [here](https://github.com/lucydot/effmass/blob/9280b9ce32e80dc849364097cca24cc1308e7b8e/effmass/analysis.py#L275). I have also changed the attribute name, from `Segment.ptype` (options: "electron" or "hole") to `Segment.band_type` (options: "conduction_band" or "valence_band") as I think the latter is easier to interpret.
I don't want effmass to make any assumptions about partial occupancies - I think it's safer if the user has to manually set whether a band is valence or conduction.
