### 2018-07-24 lucydot/effmass#7 @lucydot
I've made a couple of changes to the `effmass.inputs` module and the latest commit works for me on the data you have sent across. The read in errors were due to having two k-point blocks (one for spin up, one for spin down) in the PROCAR file when ISPIN=2. I should write some tests for spin-polarized calculations (SoC calculations are already supported and tested), but I will leave that for a later date.
