### 2018-07-24 #797 @lucydot
Thanks for the feedback @mkhorton and @ajjackson, it is really useful. @ajjackson I am chuffed you are planning to make use of it :smiley:

@mkhorton thanks for pointing out the possible testing pitfall,  I've raised it as an issue on the repo.

**Functionality**:

- I have looked at the issues raised @ajjackson ([here](https://github.com/lucydot/effmass/issues/5) and [here](https://github.com/lucydot/effmass/issues/7)), let me know whether these solve the problem and if so I will close them.

**Documentation**:

- I have fixed the issue regarding out of date docstring [here](https://github.com/lucydot/effmass/issues/6).
- I have added some module level docstrings to the API documentation [here](https://effmass.readthedocs.io/en/latest/API%20documentation.html).

**References**

- I have included academic references for pymatgen. I will add a reference for sumo once it has a Zenodo / JOSS DOI.
- I haven't included the Seebeck reference because, as you say, it isn't implemented in effmass. I do include a link to a paper which contains a definition and further references.
