### 2018-07-24 #797 @katyhuff
This looks like a very comprehensive response @lucydot . I think that these changes are sufficient and hopefully @ajjackson will confirm. Assuming this is confirmed, I think this is ready to accept. I know you intend to include a reference to the sumo JOSS paper, but I don't think the review process for sumo should delay the review process for this work. If a reference to the sumo repository can be added, perhaps that is a good middle ground.

In any case, as soon as we get confirmation from @ajjackson that these issues satisfy the concerns, this will be ready to accept. At that point could you please make an archive of the reviewed software in Zenodo/figshare/other service and update this thread with the DOI of the archive? I can then move forward with accepting the submission.
