### 2018-07-24 lucydot/effmass#5 @ajjackson
That's great, makes it easy to fix up an ambiguous case!

I'm still having a bit of trouble reading the data in though. Here's my data: a PBE calculation on CdTe with a very small gap, and some broadening across the gap. When I try to generate_segments with modest settings it fails to find any of the valence band states. I can get them to appear if I set a large enough _extrema_search_depth_, but this only works if the value of _energy_range_ is just as large and that seems less than ideal as it can affect accuracy.

I tried moving the Fermi energy in the Data object to help with searching in both directions, but this mostly resulted in generate_segments failing to find any segments at all.

[partial_occ.zip](https://github.com/lucydot/effmass/files/2225164/partial_occ.zip)
