### 2018-07-24 lucydot/effmass#7 @ajjackson
This is reading in for me now, but I'm not sure the results are consistent?

Here's a non-spin-polarised (closed-shell) calculation on the same system.  [test_mgo_nospin.zip](https://github.com/lucydot/effmass/files/2225297/test_mgo_nospin.zip)
(The setup was somewhat sloppy, so small differences are to be expected.)

Over the segment 5.33 eV [0. 0. 0.] -> [0.05 0.05 0.05] the _effmass_ methods all give me an effmass around 0.37, while in the spin-polarised case over the segment 5.00 eV [0. 0. 0.] -> [0.025 0.025 0.025] effmass is giving me 0.089 with all three methods.
(The last line of effmass_test.py needs to be changed to `display_effmasses(segments[-7])` to select this segment.)

By comparison, sumo-bandstats gives an effective mass of 0.375 for G->L in the closed-shell calculation and 0.350 for G->L in the open-shell calculation.
