### 2018-07-28 d83e8fe
fixed workflow for partial occupancies
