### 2018-07-28 7d4eee9
remove redundant test
