### 2018-07-28 bfff260
improve warning message
