+++ effmass/extrema.py
@@ -25,7 +25,7 @@ def _check_partial_occupancy(occupancy):
     """
     if np.all(np.in1d(occupancy,([0, 1, 2]))) is not True:
         warnings.warn(
-            "You have partial occupation numbers in Data.occupancy. Any eigenstate with an occupancy > 0.5 will be considered occupied. You may wish to set the attributes Data.VBM, Data.CBM and Data.fermi_energy manually."
+            "You have partial occupation numbers in Data.occupancy. Any eigenstate with an occupancy > 0.5 will be considered occupied. You should check that the attributes Data.VBM, Data.CBM and Data.fermi_energy are correct, and if not, set them manually."
         )
 
 
