### 2018-07-28 f748777
fix test for whether there are partial occupancies or not
