### 2018-07-28 cd55ea5
improve error message
