### 2018-07-28 lucydot/effmass#5 @lucydot
I've improved the warning messages for partial occupancies - the key is to adjust the data.CBM and data.VBM, as these are determined from the occupancy.

This is the work flow I have in mind (I have tested for the data above, and it works for me with the latest commit to master).
1. instantiate settings object
2. instantiate data object. If there are partial occupancies, there is the following warning
>"UserWarning: You have partial occupation numbers in Data.occupancy. You should check that the attributes Data.VBM, Data.CBM and Data.fermi_energy are correct, and if not, set them manually."
3. If Data.VBM or Data.CBM are incorrect (they will be incorrect when occupancy < 0.5 in valence band, and > 0.5 in conduction band) then set `Data.VBM`, `Data.CBM`, `Data.fermi_energy` manually (it is fairly easy to find out what the correct values are from `Data.energies` and `Data.occupancy`).
4. generate segments.

I've set the cutoff for partial occupancies at 0.5: anything above 0.5 is considered occupied, anything below 0.5 is considered unoccupied. For your test case, there was occupancy > 0.5 in the conduction band which was causing the problems. I _think_ it is rare to have partial occupancy this extreme, although I may well be wrong. If the partial occupancy is below 0.5 in the conduction band, then the warning method will still print, but the `Segment.CBM`, `Segment.VBM` or  `Segment.fermi_energy` does not need to be manually changed.

Perhaps I need to write a mini tutorial for this.
