### 2018-07-29 #797 @katyhuff
It looks like there are still two open issues that you're communicating in ([here](https://github.com/lucydot/effmass/issues/5) and [here](https://github.com/lucydot/effmass/issues/7]). Perhaps one of you (@lucydot @ajjackson) when we're ready to move forward with the submission.
