### 2018-07-30 #797 @lucydot
Yes, I'd like to sort out the discrepancy with the sumo package in [issue #7](https://github.com/lucydot/effmass/issues/7)  - it looks like effmass is not handling spin-polarised data properly.
