### 2018-07-30 lucydot/effmass#5 @ajjackson
That works for me, and seems a good solution to the problem!
