### 2018-07-30 #797 @ajjackson
Yes, [Issue 5 (partial occupancy)](https://github.com/lucydot/effmass/issues/5) is fine now. Just waiting on [issue 7](https://github.com/lucydot/effmass/issues/7) which affects the calculated values.
