### 2018-07-30 e88ec88
fixed read in of energies and occupancies for spin polarised data
