### 2018-07-30 229b065
fix bug when reading in and ignoring kpoints at the start of the file
