### 2018-07-30 lucydot/effmass#7 @lucydot
Thanks for finding this problem @ajjackson - the latest commit on master has fixed it and results for the spin-polarised case are now consistent with Sumo (0.35 calculated by all three methods for 5.00 eV [0. 0. 0.] -> [0.025 0.025 0.025]). The problem arose partly because the data structures I am using are not ideal - a task on the todo list is to store the data in a pandas array rather than numpy arrays.
