### 2018-07-30 #797 @lucydot
Latest commit to master fixes [issue 7](https://github.com/lucydot/effmass/issues/7)- thank-you for finding this bug @ajjackson, it was an important one!
