### 2018-07-30 lucydot/effmass#7 @ajjackson
Great! That case looks good now.

I'm having trouble loading other cases though, with errors like:
```
Traceback (most recent call last):
  File "./effmass_test.py", line 10, in <module>
    data = effmass.inputs.Data("OUTCAR", "PROCAR")
  File "/home/adam/src/effmass/effmass/inputs.py", line 117, in __init__
    self.number_of_bands).T[i][:284]
ValueError: could not broadcast input array from shape (284) into shape (318)
```

That hard-coded "284" in _effmass.inputs_ looks suspicious, it's the same as the number of k-points in the test case!
