### 2018-07-30 lucydot/effmass#7 @lucydot
Ah yes, and it slipped through testing as I've not any spin-polarised test cases yet - I'll fix it now.
