### 2018-07-30 79c173c
removed hardcoded number of kpoints
