### 2018-07-30 lucydot/effmass#7 @lucydot
I have removed the hard coded value for number of kpoints, and this latest commit works for me now on a different test calculation.
