### 2018-07-31 lucydot/effmass#7 @ajjackson
I'm running into e.g.
```
IndexError: index 24 is out of bounds for axis 0 with size 24
```
when trying to instantiate `Data`

[ge-band-squash-split.zip](https://github.com/lucydot/effmass/files/2245211/ge-band-squash-split.zip)
