### 2018-08-01 lucydot/effmass#7 @lucydot
This needs a test to catch mistakes - are you ok with me using the latest spin-polarised data you have posted as a test case?
