### 2018-08-01 0bb3178
fixed bug for spin polarised read in and wrote test case for SPusing data from Adam Jackson
