+++ effmass/inputs.py
@@ -115,7 +115,7 @@ class Data():
                 energies[i] = vasp_data.bands[:, 1:].reshape(
                                             number_of_kpoints*2,  # factor or 2 for each kpoint block
                                             self.number_of_bands).T[i][:number_of_kpoints]
-                energies[20+i] = vasp_data.bands[:, 1:].reshape(
+                energies[self.number_of_bands+i] = vasp_data.bands[:, 1:].reshape(
                                             number_of_kpoints*2,
                                             self.number_of_bands).T[i][number_of_kpoints:]
             occupancy = np.zeros([self.number_of_bands*2,number_of_kpoints])
@@ -123,7 +123,7 @@ class Data():
                 occupancy[i] = vasp_data.occupancy[:, 1:].reshape(
                                                  number_of_kpoints*2,
                                                  self.number_of_bands).T[i][:number_of_kpoints]
-                occupancy[20+i] = vasp_data.occupancy[:, 1:].reshape(
+                occupancy[self.number_of_bands+i] = vasp_data.occupancy[:, 1:].reshape(
                                                  number_of_kpoints*2,
                                                  self.number_of_bands).T[i][number_of_kpoints:]
         else:
