### 2018-08-01 919d26c
fixed bug for spin polarised read in and wrote test case for SPusing data from Adam Jackson
