### 2018-08-01 lucydot/effmass#7 @lucydot
Ok - with the latest commit to master you should be able to instantiate the Data class for spin-polarised cases. Also, Travis will  automatically test effmass for the Ge spin-polarised data.
