### 2018-08-01 #797 @lucydot
Issue 7 is not quite closed (but very close!), when it is I will archive and update this thread.
