### 2018-08-03 lucydot/effmass#7 @ajjackson
Looks good to me! It seems like a good idea to have test cases for all the various spin/SOC combinations.
