### 2018-08-03 #797 @ajjackson
Issue 7 looks good now, I have ticked off "functionality".
