### 2018-08-03 #797 @katyhuff
Excellent. @lucydot when you make archive of the reviewed software in Zenodo/figshare/other service, please just update this thread with the DOI of the archive and we'll move forward with acceptance! (You may want to be sure to double check the paper, all minor details, etc. before creating the archive).
