### 2018-08-06 d68db96
Update paper.md
