### 2018-08-06 2e63324
Update paper.bib
