+++ paper.bib
@@ -57,9 +57,8 @@ volume = "68",
 pages = "314 - 319",
 year = "2013",
 issn = "0927-0256",
-doi = "https://doi.org/10.1016/j.commatsci.2012.10.028",
-url = "http://www.sciencedirect.com/science/article/pii/S0927025612006295",
-author = "Shyue Ping Ong and William Davidson Richards and Anubhav Jain and Geoffroy Hautier and Michael Kocher and Shreyas Cholia and Dan Gunter and Vincent L. Chevrier and Kristin A. Persson and Gerbrand Ceder",
+url = "https://doi.org/10.1016/j.commatsci.2012.10.028"
+author = "S. P. Ong and W. D. Richards and A. Jain and G. Hautier and M. Kocher and S. Cholia and D. Gunter and V. L. Chevrier and K. A. Persson and G. Ceder",
 keywords = "Materials, Project, Design, Thermodynamics, High-throughput"
 }
 	
