### 2018-08-06 d0830a4
Update paper.bib
