+++ paper.bib
@@ -57,7 +57,7 @@ volume = "68",
 pages = "314 - 319",
 year = "2013",
 issn = "0927-0256",
-url = "https://doi.org/10.1016/j.commatsci.2012.10.028"
+url = "https://doi.org/10.1016/j.commatsci.2012.10.028",
 author = "S. P. Ong and W. D. Richards and A. Jain and G. Hautier and M. Kocher and S. Cholia and D. Gunter and V. L. Chevrier and K. A. Persson and G. Ceder",
 keywords = "Materials, Project, Design, Thermodynamics, High-throughput"
 }
