### 2018-08-06 #797 @lucydot
Hi @katyhuff  - here is the DOI: 10.5281/zenodo.1338919
Thank-you for guiding the review process, `effmass` gained much from it.
